test: 페이크 클래스들 한 패키지안으로 이동
