+++ ark-di/arik-di-android/src/test/java/com/re4rk/arkdiAndroid/fakeClasses/FakeActivity.kt
@@ -1,0 +1,5 @@
+package com.re4rk.arkdiAndroid.fakeClasses
+
+import com.re4rk.arkdiAndroid.ArkAppCompatActivity
+
+class FakeActivity : ArkAppCompatActivity()
