+++ ark-di/arik-di-android/src/test/java/com/re4rk/arkdiAndroid/fakeClasses/fakeDependency/FakeActivityDependency.kt
@@ -1,0 +1,3 @@
+package com.re4rk.arkdiAndroid.fakeClasses.fakeDependency
+
+interface FakeActivityDependency
