+++ ark-di/arik-di-android/src/test/java/com/re4rk/arkdiAndroid/fakeClasses/fakeDependency/FakeRetainedActivityDependencyImpl.kt
@@ -1,3 +1,3 @@
-package com.re4rk.arkdiAndroid.fakeDependency
+package com.re4rk.arkdiAndroid.fakeClasses.fakeDependency
 
 class FakeRetainedActivityDependencyImpl : FakeRetainedActivityDependency
