+++ ark-di/arik-di-android/src/test/java/com/re4rk/arkdiAndroid/fakeClasses/fakeModules/FakeActivityModule.kt
@@ -1,9 +1,9 @@
-package com.re4rk.arkdiAndroid.fakeModules
+package com.re4rk.arkdiAndroid.fakeClasses.fakeModules
 
 import android.content.Context
 import com.re4rk.arkdi.ArkModule
-import com.re4rk.arkdiAndroid.fakeDependency.FakeActivityDependency
-import com.re4rk.arkdiAndroid.fakeDependency.FakeActivityDependencyImpl
+import com.re4rk.arkdiAndroid.fakeClasses.fakeDependency.FakeActivityDependency
+import com.re4rk.arkdiAndroid.fakeClasses.fakeDependency.FakeActivityDependencyImpl
 
 class FakeActivityModule(parentModule: ArkModule, context: Context) : ArkModule(parentModule) {
     fun fakeActivityDependency(): FakeActivityDependency = FakeActivityDependencyImpl()
