+++ ark-di/arik-di-android/src/test/java/com/re4rk/arkdiAndroid/fakeClasses/fakeModules/FakeApplicationModule.kt
@@ -1,9 +1,9 @@
-package com.re4rk.arkdiAndroid.fakeModules
+package com.re4rk.arkdiAndroid.fakeClasses.fakeModules
 
 import android.content.Context
 import com.re4rk.arkdi.ArkModule
-import com.re4rk.arkdiAndroid.fakeDependency.FakeServiceDependency
-import com.re4rk.arkdiAndroid.fakeDependency.FakeServiceDependencyImpl
+import com.re4rk.arkdiAndroid.fakeClasses.fakeDependency.FakeServiceDependency
+import com.re4rk.arkdiAndroid.fakeClasses.fakeDependency.FakeServiceDependencyImpl
 
 class FakeApplicationModule(context: Context) : ArkModule() {
     fun fakeApplicationDependency(): FakeServiceDependency = FakeServiceDependencyImpl()
