+++ ark-di/arik-di-android/src/test/java/com/re4rk/arkdiAndroid/fakeClasses/fakeModules/FakeRetainedActivityModule.kt
@@ -1,9 +1,9 @@
-package com.re4rk.arkdiAndroid.fakeModules
+package com.re4rk.arkdiAndroid.fakeClasses.fakeModules
 
 import android.content.Context
 import com.re4rk.arkdi.ArkModule
-import com.re4rk.arkdiAndroid.fakeDependency.FakeRetainedActivityDependency
-import com.re4rk.arkdiAndroid.fakeDependency.FakeRetainedActivityDependencyImpl
+import com.re4rk.arkdiAndroid.fakeClasses.fakeDependency.FakeRetainedActivityDependency
+import com.re4rk.arkdiAndroid.fakeClasses.fakeDependency.FakeRetainedActivityDependencyImpl
 
 class FakeRetainedActivityModule(parentModule: ArkModule, context: Context) : ArkModule(parentModule) {
     fun fakeRetainedActivityDependency(): FakeRetainedActivityDependency =
