--- ark-di/arik-di-android/src/test/java/com/re4rk/arkdiAndroid/fakeDependency/FakeRetainedActivityDependency.kt
@@ -1,3 +1,0 @@
-package com.re4rk.arkdiAndroid.fakeDependency
-
-interface FakeRetainedActivityDependency
